pyproject.toml
[tool.poetry]
name = "box-office-tracking"
version = "0.1.0"
description = ""
authors = ["Ethan Fuerst"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gspread = "^5.12.4"
pandas = "^2.1.4"
gspread-formatting = "^1.1.2"
duckdb = "^1.1.3"
modal = "^0.64.52"
python-dotenv = "^1.0.1"
sqlfluff = "^3.3.0"
lxml = "^5.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = true

[tool.sqlfluff]
dialect = "duckdb"
max_line_length = 80
indent_unit = "tab"
tab_space_size = 4
capitalization_policy = "lower"
extended_capitalisation_policy = "lower"
preferred_type_casting_style = "shorthand"
exclude_rules = "LT05, LT09, ST05"

[tool.sqlfluff.layout.type.comma]
line_position = "leading"

[tool.sqlfluff.rules.layout.long_lines]
max_line_length = 80

[tool.sqlfluff.rules.layout.select_targets]
line_position = "leading"

[tool.sqlfluff.rules.structure.subquery]
forbid_subquery_in = "both"
